FIX-322 - Fix server config being exposed. Move webpack, configs and tests into internals folder.
