+++ internals/config/public.js
@@ -1,0 +1,4 @@
+const { PORT = 3000 } = process.env
+
+export const port = parseInt(PORT, 10)
+export const apiPrefix = '/api'
