+++ internals/test/helpers/mount.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { mount } from 'enzyme'
 
 import createFlux from 'flux/createFlux'
-import ApiClient from '../../shared/api-client'
+import ApiClient from '../../../shared/api-client'
 
 import I18nContainer from 'utils/i18n-container'
