+++ internals/webpack/base.config.js
@@ -11,7 +11,7 @@ export default {
     app: './app/index.js'
   },
   output: {
-    path: path.resolve(__dirname, '../dist'),
+    path: path.resolve(__dirname, '../../dist'),
     filename: '[name]-[hash].js',
     chunkFilename: '[name]-[hash].js',
     publicPath: '/assets/'
